#209 move search js files to its own tools folder
