+++ tools/search/test.js
